razer_common: only create sysfs files if all goes well
If parsing the device does not work correctly, the sysfs files will
remain, which is not ok.  Also, the devdata needs to be set _before_ the
device files are created, otherwise the file could be present and
accessed before the pointer is set, causing the kernel to crash.

Fix all of this by just moving the creation of the files to the end of
the function.

This is still racy, and incorrect (no error checking happens), but the
correct fix for this will only work on newer kernels and I do not know
if the driver is ready to be tied to only modern kernel versions or not.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>